requirements.txt
allure-pytest==2.8.12
allure-python-commons==2.8.12
asgiref
atomicwrites==1.3.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
idna==2.10
importlib-metadata==1.6.0
iniconfig==1.0.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.2
pyparsing==2.4.7
pytest==7.0.1
pytest-assume==2.2.1
pytest-ordering==0.6
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.1.0
Django
pytest-xdist
pytest-parallel
gunicorn
pytest-htmjsonpath~=0.82